move user interface globals to the correct object files
- improves build time since we don't need to include everything for the globals in every file that includes the user interface code
