+++ xlive/Blam/Engine/interface/screens/screen_4way_signin.cpp
@@ -12,14 +12,13 @@
 #include "interface/user_interface.h"
 #include "interface/user_interface_bitmap_block.h"
 #include "interface/user_interface_controller.h"
-#include "interface/user_interface_globals.h"
 #include "interface/user_interface_memory.h"
 #include "interface/user_interface_model_block.h"
 #include "interface/user_interface_networking.h"
 #include "interface/user_interface_player_widget.h"
 #include "interface/user_interface_screen_widget_definition.h"
 #include "interface/user_interface_utilities.h"
-#include "Networking/online/online_account_xbox.h"
+#include "networking/online/online_account_xbox.h"
 #include "shell/shell.h"
 #include "tag_files/global_string_ids.h"
 #include "text/unicode.h"
