+++ xlive/Blam/Engine/interface/screens/screen_brightness_menu.h
@@ -1,5 +1,4 @@
 #pragma once
-
 #include "interface/user_interface_widget.h"
 #include "interface/user_interface_widget_list.h"
 #include "interface/user_interface_widget_list_item.h"
