+++ xlive/Blam/Engine/interface/screens/screen_controller_settings_menu.cpp
@@ -2,7 +2,6 @@
 #include "screen_controller_settings_menu.h"
 
 #include "input/input_abstraction.h"
-#include "interface/user_interface_globals.h"
 #include "interface/user_interface_widget_list.h"
 #include "interface/user_interface_widget_list_item.h"
 #include "interface/user_interface_widget_text.h"
