+++ xlive/Blam/Engine/interface/user_interface_controller.cpp
@@ -2,7 +2,6 @@
 #include "user_interface_controller.h"
 
 #include "user_interface.h"
-#include "user_interface_globals.h"
 #include "user_interface_utilities.h"
 #include "user_interface_widget_window.h"
 #include "user_interface_guide.h"
