+++ xlive/Blam/Engine/interface/user_interface_widget_window.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
-
 #include "user_interface_widget_window.h"
+
 #include "rasterizer/rasterizer_globals.h"
 
 c_screen_widget::c_screen_widget(e_user_interface_screen_id menu_id, e_user_interface_channel_type channel_type, e_user_interface_render_window window_index, uint16 user_flags) :
@@ -43,7 +43,7 @@ c_screen_widget::c_screen_widget(e_user_interface_screen_id menu_id, e_user_inte
 	this->field_A01 = false;
 }
 
-const e_user_interface_screen_id c_screen_widget::get_id() const
+e_user_interface_screen_id c_screen_widget::get_id() const
 {
 	return m_screen_id;
 }
